hwdb: Add support for the buttons on CZC P10T tablet
This requires the tablet to be in "Android" mode. In default ("Windows 7")
mode, the Home button emits multiple key presses for a Windows-specific key
combination while the second button toggles the RF Kill Switch in addition
to producing a scancode

There are two varitions of the machine that differ by the label on the
second button.